- call stack matching for exceptional flows - new table for archive file and log file list and status - 2 new methods for fetching data for the tree - use long everywhere (insight from jmc) - tree model is work in progress, need to show test candidates when a method node is selected - create flow for creating database for a new session - process all threads discovered when scanning log files

+++ src/main/java/com/insidious/plugin/pojo/MethodCallExpression.java
@@ -239,7 +239,7 @@ public class MethodCallExpression implements Expression {
             }
 
 
-            Parameter existingVariableById = variableContainer.getParametersById(value);
+            Parameter existingVariableById = variableContainer.getParametersById((long) value);
             if (existingVariableById != null) {
                 if (overrideName && !returnValue.getName().equals(existingVariableById.getName())) {
                     returnValue.setName(existingVariableById.getName());
